requirements.txt
litellm>=1.40.0            
python-dotenv>=1.0.0       

aiohttp>=3.9.0          
tenacity>=8.2.0        

structlog>=24.1.0    

asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
pgvector>=0.2.0
rank-bm25>=0.2.2

mcp>=1.0.0
httpx>=0.27.0